change precision from float to double

+++ src/main/java/me/chriss99/Main.java
@@ -135,7 +135,7 @@ public class Main {
             }
         }
 
-        float[][] heightMap = VAOGenerator.pillar(100, 100);
+        double[][] heightMap = VAOGenerator.pillar(100, 100);
 
         vaoList.add(VAOGenerator.heightMapToSimpleVAO(heightMap));
     }
